requirements.txt
qrcode[pil]>=7.4
Pillow>=9.0
pytest>=7.0
streamlit>=1.20
